requirements.txt
streamlit==0.82.0
pandas==1.0.5
nltk==3.6.2
numpy==1.18.5
Pillow==8.2.0
scikit_learn==0.24.2
